requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.0.76
Pillow==10.0.0
